requirements.txt
langchain
langchain-community
langchain-openai
langchain-huggingface
langchain-pinecone
pinecone
sentence-transformers
streamlit
python-dotenv
pypdf
